+++ src/main/java/bookclub/repositories/BookRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
     List<Book> findByUser(User user);
 
-    List<Book> findByUserIdIn(List<Integer> userIds);
+    List<Book> findByUserIdIn(List<Long> userIds);
 }
